requirements.txt
ffmpy
pyTelegramBotAPI
python_twitter
